Fix leader election config and TLS config loading of partial values.
Fixes #1367

Also add some tests for the YAML parsing of the config file.
